blender/requirements.txt
pillow~=11.1
rich~=14.0
xxhash~=3.5
